+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2024-08-17
+### Fixed
+- FilenotFoundError when trying to remove file from local storage if it is not exists. Added flag to skip on error.
+
+## [1.6.0] - 2024-08-17
+### Fixed
+ - Local media manager not removing files.
+### Removed
+- Deta storage repo code and tests.
+- pdbpp from dev dependencies.
+### Changed
+- Updated yt-dlp from 2024.7.1 to version 2024.8.6.
+
+## [1.5.2] - 2024-07-05
+### Changed
+- Bumped up version of yt-dlp from 2024.5.27 to 2024.7.1.
+
 ## [1.5.1] - 2024-06-22
 ### Changed
 - Bumped up version of yt-dlp from 2023.12.30 to 2024.5.27.